close #350, (partial, not models yet) seperation of ok_tag app from auxiliary. Seperation of mixins in main tag view, to allow more movement to get more links behavior, starting to refactor tests, seperation of knesset selector in tags view to use proper GET param and not 'page' which is quite misleading, adding order by date to meetings in tag
