+++ auxiliary/models.py
@@ -1,15 +1,13 @@
-import re
-from django.db import models
-from django.utils.translation import ugettext_lazy as _
-from tinymce import models as tinymce_models
 from django.contrib.auth.models import User
-from django.contrib.contenttypes.models import ContentType
-from django.db.models.signals import post_save, post_delete
-from tagging.models import TaggedItem, Tag
-from committees.models import CommitteeMeeting
 from django.contrib.contenttypes import generic
-from knesset.utils import trans_clean
+from django.contrib.contenttypes.models import ContentType
+from django.db import models
 
+from django.utils.translation import ugettext_lazy as _
+from tagging.models import Tag
+from tinymce import models as tinymce_models
+
+from committees.models import CommitteeMeeting
 
 ICON_CHOICES = (
     ('quote', _('Quote')),
@@ -19,7 +17,6 @@ ICON_CHOICES = (
 
 
 class TidbitManager(models.Manager):
-
     def get_query_set(self):
         return super(TidbitManager, self).get_query_set().filter(
             is_active=True).order_by('ordering')
@@ -93,119 +90,6 @@ class TagSuggestion(models.Model):
     object = generic.GenericForeignKey('content_type', 'object_id')
 
 
-def add_tags_to_related_objects(sender, instance, **kwargs):
-    """
-    When a tag is added to an object, we also tag other objects that are
-    related.
-    This currently only handles tagging of bills. When a bill is tagged it will
-    tag related votes and related committee meetings.
-
-    """
-    obj = instance.object
-    tag = instance.tag
-    bill_ctype = ContentType.objects.get(app_label='laws', model='bill')
-    if type(obj) is bill_ctype.model_class():
-        vote_ctype = ContentType.objects.get(app_label='laws', model='vote')
-        # tag related votes
-        for v in obj.pre_votes.all():
-            (ti, created) = TaggedItem._default_manager.get_or_create(
-                tag=tag,
-                content_type=vote_ctype,
-                object_id=v.id)
-        v = obj.first_vote
-        if v:
-            (ti, created) = TaggedItem._default_manager.get_or_create(
-                tag=tag,
-                content_type=vote_ctype,
-                object_id=v.id)
-        v = obj.approval_vote
-        if v:
-            (ti, created) = TaggedItem._default_manager.get_or_create(
-                tag=tag,
-                content_type=vote_ctype,
-                object_id=v.id)
-
-        cm_ctype = ContentType.objects.get_for_model(CommitteeMeeting)
-        for cm in obj.first_committee_meetings.all():
-            (ti, created) = TaggedItem._default_manager.get_or_create(
-                tag=tag,
-                content_type=cm_ctype,
-                object_id=cm.id)
-        for cm in obj.second_committee_meetings.all():
-            (ti, created) = TaggedItem._default_manager.get_or_create(
-                tag=tag,
-                content_type=cm_ctype,
-                object_id=cm.id)
-
-
-post_save.connect(add_tags_to_related_objects, sender=TaggedItem)
-
-
-def remove_tags_from_related_objects(sender, instance, **kwargs):
-    obj = instance.object
-    try:
-        tag = instance.tag
-    except Tag.DoesNotExist:  # the tag itself was deleted,
-        return  # so we have nothing to do.
-    bill_ctype = ContentType.objects.get(app_label='laws', model='bill')
-    if type(obj) is bill_ctype.model_class():
-        vote_ctype = ContentType.objects.get(app_label='laws', model='vote')
-        # untag related votes
-        for v in obj.pre_votes.all():
-            try:
-                ti = TaggedItem._default_manager.get(
-                    tag=tag,
-                    content_type=vote_ctype,
-                    object_id=v.id)
-                ti.delete()
-            except TaggedItem.DoesNotExist:
-                pass
-        v = obj.first_vote
-        if v:
-            try:
-                ti = TaggedItem._default_manager.get(
-                    tag=tag,
-                    content_type=vote_ctype,
-                    object_id=v.id)
-                ti.delete()
-            except TaggedItem.DoesNotExist:
-                pass
-        v = obj.approval_vote
-        if v:
-            try:
-                ti = TaggedItem._default_manager.get(
-                    tag=tag,
-                    content_type=vote_ctype,
-                    object_id=v.id)
-                ti.delete()
-            except TaggedItem.DoesNotExist:
-                pass
-
-        # untag related committee meetings
-        cm_ctype = ContentType.objects.get_for_model(CommitteeMeeting)
-        for cm in obj.first_committee_meetings.all():
-            try:
-                ti = TaggedItem._default_manager.get(
-                    tag=tag,
-                    content_type=cm_ctype,
-                    object_id=cm.id)
-                ti.delete()
-            except TaggedItem.DoesNotExist:
-                pass
-        for cm in obj.second_committee_meetings.all():
-            try:
-                ti = TaggedItem._default_manager.get(
-                    tag=tag,
-                    content_type=cm_ctype,
-                    object_id=cm.id)
-                ti.delete()
-            except TaggedItem.DoesNotExist:
-                pass
-
-
-post_delete.connect(remove_tags_from_related_objects, sender=TaggedItem)
-
-
 class TagKeyphrase(models.Model):
     tag = models.ForeignKey('tagging.Tag')
     phrase = models.CharField(max_length=100, blank=False, null=False)
@@ -213,25 +97,11 @@ class TagKeyphrase(models.Model):
     def __unicode__(self):
         return u"%s - %s" % (self.tag, self.phrase)
 
-
-def tag_vote(vote):
-    vote_ctype = ContentType.objects.get(app_label='laws', model='vote')
-    t = vote.title.translate(trans_clean)
-    t = re.sub(' . ', ' ', t)
-    t = re.sub(' +', ' ', t)
-    for tag_phrase in TagKeyphrase.objects.all().select_related('tag'):
-        if tag_phrase.phrase in t:
-            TaggedItem.objects.get_or_create(tag=tag_phrase.tag,
-                                             content_type=vote_ctype,
-                                             object_id=vote.id)
-
-
-
 # The following commented code is a part of an attempt to auto-tag
 # that wasn't successful enough, mostly because a lot of interesting
 # tags don't have enough training data.
 
-#def create_tokens(list_of_strings):
+# def create_tokens(list_of_strings):
 #    tokens = defaultdict(lambda: 0)
 #    for s in list_of_strings:
 #        t = s.translate(trans_dict)
@@ -244,7 +114,7 @@ def tag_vote(vote):
 #    return tokens
 #
 #
-#def tag_common_tokens(tag):
+# def tag_common_tokens(tag):
 #    t = Vote.objects.filter(tagged_items__tag=tag).values_list('title', flat=True)
 #    t = list(set(t))
 #    l = len(t)
@@ -255,7 +125,7 @@ def tag_vote(vote):
 #    return a
 #
 #
-#def create_tag_keyphrases(queryset, field_name):
+# def create_tag_keyphrases(queryset, field_name):
 #    skip_tokens = set()  # used to speed up create_tag_keyphrases
 #    ct = ContentType.objects.get_for_model(queryset[:1][0])
 #    for o in queryset:
@@ -294,20 +164,3 @@ def tag_vote(vote):
 #                    print token.encode('utf8')
 #                    print t, t in o.tags
 #                    print token_objs
-
-
-def tagged_votes_titles(tags):
-    """returns a list representation of the titles (cleaned) of votes tagged
-       by the given tags. There are also lines for the tags themselves as
-       headers. Use the output to write to file"""
-    res = []
-    for tag in tags:
-        res.append("\n%s" % tag.name)
-        t = Vote.objects.filter(tagged_items__tag=tag).values_list('title', flat=True)
-        t = list(set(t))
-        for t0 in t:
-            t1 = t0.translate(trans_clean)
-            t1 = re.sub(' . ', ' ', t1)  # remove single char 'words'
-            t1 = re.sub(' +', ' ', t1)  # unify blocks of spaces
-            res.append(t1)
-    return res
