+++ auxiliary/views.py
@@ -1,99 +1,29 @@
-import csv, random, tagging, logging
-import json
-from actstream import action
+import logging
+import random
+from datetime import timedelta
+
 from annotatetext.views import post_annotation as annotatetext_post_annotation
 from django.conf import settings
-from django.contrib.auth.decorators import login_required, permission_required
 from django.contrib.comments.models import Comment
 from django.contrib.contenttypes.models import ContentType
 from django.core.cache import cache
-from django.core.urlresolvers import reverse
-from django.db.models import Q
 from django.http import (
-    HttpResponseForbidden, HttpResponseRedirect, HttpResponse,
-    HttpResponseNotAllowed, HttpResponseBadRequest, Http404, HttpResponsePermanentRedirect)
+    HttpResponseForbidden, HttpResponseRedirect)
 from django.shortcuts import render_to_response, get_object_or_404
 from django.template import RequestContext
-from django.utils.translation import ugettext as _
-from django.views.generic import TemplateView, DetailView, ListView
-from django.views.generic.list import BaseListView
-from django.views.decorators.http import require_http_methods
-from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
-from tagging.models import Tag, TaggedItem
 from django.utils import timezone
-from datetime import timedelta
+from django.utils.translation import ugettext as _
+from django.views.decorators.http import require_http_methods
+from django.views.generic import TemplateView, DetailView, ListView
+from okscraper_django.models import ScraperRun
 
-from auxiliary import utils
-from .forms import TidbitSuggestionForm, FeedbackSuggestionForm, TagSuggestionForm
-from .models import Tidbit, TagSuggestion
 from committees.models import CommitteeMeeting
 from events.models import Event
-from knesset.utils import notify_responsible_adult
 from laws.models import Vote, Bill
-from mks.models import Member, Knesset
-from tagging.utils import get_tag
-from auxiliary.models import TagSynonym
+from mks.models import Member
 
-from okscraper_django.models import ScraperRun, ScraperRunLog
-
-
-class BaseTagMemberListView(ListView):
-    """Generic helper for common tagged objects and optionally member
-    operations. Shoud be inherited by others"""
-
-    url_to_reverse = None  # override in inherited for reversing tag_url
-
-    # in context
-
-    @property
-    def tag_instance(self):
-        if not hasattr(self, '_tag_instance'):
-            tag = self.kwargs['tag']
-            self._tag_instance = get_tag(tag)
-
-            if self._tag_instance is None:
-                raise Http404(_('No Tag found matching "%s".') % tag)
-
-        return self._tag_instance
-
-    @property
-    def member(self):
-        if not hasattr(self, '_member'):
-            member_id = self.request.GET.get('member', False)
-
-            if member_id:
-                try:
-                    member_id = int(member_id)
-                except ValueError:
-                    raise Http404(
-                        _('No Member found matching "%s".') % member_id)
-
-                self._member = get_object_or_404(Member, pk=member_id)
-            else:
-                self._member = None
-
-        return self._member
-
-    def get_context_data(self, *args, **kwargs):
-        context = super(BaseTagMemberListView, self).get_context_data(
-            *args, **kwargs)
-
-        context['tag'] = self.tag_instance
-        context['tag_url'] = reverse(self.url_to_reverse,
-                                     args=[self.tag_instance])
-
-        if self.member:
-            context['member'] = self.member
-            context['member_url'] = reverse(
-                'member-detail', args=[self.member.pk])
-
-        user = self.request.user
-        if user.is_authenticated():
-            context['watched_members'] = user.get_profile().members
-        else:
-            context['watched_members'] = False
-
-        return context
+from .forms import TidbitSuggestionForm, FeedbackSuggestionForm
+from .models import Tidbit
 
 
 class MainScraperStatusView(ListView):
@@ -132,6 +62,8 @@ def help_page(request):
 
         tags_cloud = cache.get('tags_cloud', None)
         if not tags_cloud:
+            # TODO: ugly hack, remove this import later, when I figure out why this is even needed here
+            from ok_tag.views import calculate_cloud_from_models
             tags_cloud = calculate_cloud_from_models(Vote, Bill, CommitteeMeeting)
             tags_cloud.sort(key=lambda x: x.name)
             cache.set('tags_cloud', tags_cloud, settings.LONG_CACHE_TIME)
@@ -270,27 +202,6 @@ def post_feedback(request):
     return form.get_response()
 
 
-@require_http_methods(['POST'])
-def suggest_tag_post(request):
-    "Post a tag suggestion form"
-    if not request.user.is_authenticated:
-        return HttpResponseForbidden()
-
-    form = TagSuggestionForm(request.POST)
-    if form.is_valid():
-        content_type = ContentType.objects.get_by_natural_key(form.cleaned_data['app_label'],
-                                                              form.cleaned_data['object_type'])
-        object = content_type.get_object_for_this_type(pk=form.cleaned_data['object_id'])
-        ts = TagSuggestion(
-            name=form.cleaned_data['name'],
-            suggested_by=request.user,
-            object=object
-        )
-        ts.save()
-
-    return form.get_response()
-
-
 def post_annotation(request):
     if request.user.has_perm('annotatetext.add_annotation'):
         return annotatetext_post_annotation(request)
@@ -354,414 +265,3 @@ class CommentsView(ListView):
     queryset = Comment.objects.order_by("-submit_date")
 
     paginate_by = 20
-
-
-def _add_tag_to_object(user, app, object_type, object_id, tag):
-    ctype = ContentType.objects.get_by_natural_key(app, object_type)
-    (ti, created) = TaggedItem._default_manager.get_or_create(
-        tag=tag,
-        content_type=ctype,
-        object_id=object_id)
-    action.send(user, verb='tagged', target=ti, description='%s' % (tag.name))
-    url = reverse('tag-detail', kwargs={'slug': tag.name})
-    return HttpResponse("{'id':%d, 'name':'%s', 'url':'%s'}" % (tag.id,
-                                                                tag.name,
-                                                                url))
-
-
-@login_required
-def add_tag_to_object(request, app, object_type, object_id):
-    """add a POSTed tag_id to object_type object_id by the current user"""
-    if request.method == 'POST' and 'tag_id' in request.POST:  # If the form has been submitted...
-        tag = get_object_or_404(Tag, pk=request.POST['tag_id'])
-        return _add_tag_to_object(request.user, app, object_type, object_id, tag)
-
-    return HttpResponseNotAllowed(['POST'])
-
-
-@login_required
-def remove_tag_from_object(request, app, object_type, object_id):
-    """remove a POSTed tag_id from object_type object_id"""
-    ctype = ContentType.objects.get_by_natural_key(app, object_type)
-    if request.method == 'POST' and 'tag_id' in request.POST:  # If the form has been submitted...
-        tag = get_object_or_404(Tag, pk=request.POST['tag_id'])
-        ti = TaggedItem._default_manager.filter(tag=tag, content_type=ctype, object_id=object_id)
-        if len(ti) == 1:
-            logger.debug('user %s is deleting tagged item %d' % (request.user.username, ti[0].id))
-            ti[0].delete()
-            action.send(request.user, verb='removed-tag', target=ti[0], description='%s' % (tag.name))
-        else:
-            logger.debug('user %s tried removing tag %d from object, but failed, because len(tagged_items)!=1' % (
-                request.user.username, tag.id))
-    return HttpResponse("{'id':%d,'name':'%s'}" % (tag.id, tag.name))
-
-
-@permission_required('tagging.add_tag')
-def create_tag_and_add_to_item(request, app, object_type, object_id):
-    """adds tag with name=request.POST['tag'] to the tag list, and tags the given object with it
-    ****
-    Currently not used anywhere, sine we don't want to allow users to add
-    more tags for now.
-    """
-    if request.method == 'POST' and 'tag' in request.POST:
-        tag = request.POST['tag'].strip()
-        msg = "user %s is creating tag %s on object_type %s and object_id %s".encode('utf8') % (
-            request.user.username, tag, object_type, object_id)
-        logger.info(msg)
-        notify_responsible_adult(msg)
-        if len(tag) < 3:
-            return HttpResponseBadRequest()
-        tags = Tag.objects.filter(name=tag)
-        if not tags:
-            try:
-                tag = Tag.objects.create(name=tag)
-            except Exception:
-                logger.warn("can't create tag %s" % tag)
-                return HttpResponseBadRequest()
-        if len(tags) == 1:
-            tag = tags[0]
-        if len(tags) > 1:
-            logger.warn("More than 1 tag: %s" % tag)
-            return HttpResponseBadRequest()
-        return _add_tag_to_object(request.user, app, object_type, object_id, tag)
-    else:
-        return HttpResponseNotAllowed(['POST'])
-
-
-@permission_required('tagging.add_tag')
-def add_tag_synonym(request, parent_tag_id, synonym_tag_id):
-    parent_tag = Tag.objects.get(pk=parent_tag_id)
-    synonym_tag = Tag.objects.get(pk=synonym_tag_id)
-    assert parent_tag.synonym_synonym_tag.count() == 0
-    assert synonym_tag != parent_tag
-    TagSynonym.objects.create(tag_id=parent_tag_id, synonym_tag_id=synonym_tag_id)
-    return HttpResponse('ok')
-
-
-def calculate_cloud_from_models(*args):
-    from tagging.models import Tag
-    cloud = Tag._default_manager.cloud_for_model(args[0])
-    for model in args[1:]:
-        for tag in Tag._default_manager.cloud_for_model(model):
-            if tag in cloud:
-                cloud[cloud.index(tag)].count += tag.count
-            else:
-                cloud.append(tag)
-    return tagging.utils.calculate_cloud(cloud)
-
-
-class TagList(ListView):
-    """Tags index view"""
-
-    model = Tag
-    template_name = 'auxiliary/tag_list.html'
-
-    def get_queryset(self):
-        return Tag.objects.all()
-
-    def get_context_data(self, **kwargs):
-        context = super(TagList, self).get_context_data(**kwargs)
-        tags_cloud = cache.get('tags_cloud', None)
-        if not tags_cloud:
-            tags_cloud = calculate_cloud_from_models(Vote, Bill, CommitteeMeeting)
-            tags_cloud.sort(key=lambda x: x.name)
-            cache.set('tags_cloud', tags_cloud, settings.LONG_CACHE_TIME)
-        context['tags_cloud'] = tags_cloud
-        return context
-
-
-class SelectorPaginator(Paginator):
-    def __init__(self, knesset_ids):
-        super(SelectorPaginator, self).__init__(sorted(list(knesset_ids)), 1)
-
-    def validate_number(self, number):
-        try:
-            number = int(number)
-        except (TypeError, ValueError):
-            raise PageNotAnInteger('That page number is not an integer')
-        if number not in self.object_list:
-            raise EmptyPage('That page contains no result')
-        return number
-
-    def page(self, number):
-        number = self.validate_number(number)
-        return self._get_page([number, ], self.object_list.index(number), self)
-
-    def _get_count(self):
-        return len(self.object_list)
-
-    def _get_num_pages(self):
-        return len(self.object_list)
-
-    def _get_page_range(self):
-        return self.object_list
-
-
-class TagDetail(DetailView):
-    """Tags index view"""
-
-    model = Tag
-    template_name = 'auxiliary/tag_detail.html'
-    slug_field = 'name'
-
-    def create_tag_cloud(self, tag, limit=30, bills=None, votes=None,
-                         cms=None):
-        """
-        Create tag could for tag <tag>. Returns only the <limit> most tagged members
-        """
-
-        try:
-            mk_limit = int(self.request.GET.get('limit', limit))
-        except ValueError:
-            mk_limit = limit
-        if bills is None:
-            bills = TaggedItem.objects.get_by_model(Bill, tag) \
-                .prefetch_related('proposers')
-        if votes is None:
-            votes = TaggedItem.objects.get_by_model(Vote, tag) \
-                .prefetch_related('votes')
-        if cms is None:
-            cms = TaggedItem.objects.get_by_model(CommitteeMeeting, tag) \
-                .prefetch_related('mks_attended')
-        mk_taggeds = [(b.proposers.all(), b.stage_date) for b in bills]
-        mk_taggeds += [(v.votes.all(), v.time.date()) for v in votes]
-        mk_taggeds += [(cm.mks_attended.all(), cm.date) for cm in cms]
-        current_k_start = Knesset.objects.current_knesset().start_date
-        d = {}
-        d_previous = {}
-        for tagged, date in mk_taggeds:
-            if date and (date > current_k_start):
-                for p in tagged:
-                    d[p] = d.get(p, 0) + 1
-            else:  # not current knesset
-                for p in tagged:
-                    d_previous[p] = d.get(p, 0) + 1
-        # now d is a dict: MK -> number of tagged in Bill, Vote and
-        # CommitteeMeeting in this tag, in the current knesset
-        # d_previous is similar, but for all non current knesset data
-        mks = dict(sorted(d.items(), lambda x, y: cmp(y[1], x[1]))[:mk_limit])
-        # Now only the most tagged are in the dict (up to the limit param)
-        for mk in mks:
-            mk.count = d[mk]
-        mks = tagging.utils.calculate_cloud(mks)
-
-        mks_previous = dict(sorted(d_previous.items(),
-                                   lambda x, y: cmp(y[1], x[1]))[:mk_limit])
-        for mk in mks_previous:
-            mk.count = d_previous[mk]
-        mks_previous = tagging.utils.calculate_cloud(mks_previous)
-        return mks, mks_previous
-
-    def get(self, *args, **kwargs):
-        tag = self.get_object()
-        ts = TagSynonym.objects.filter(synonym_tag=tag)
-        if len(ts) > 0:
-            proper = ts[0].tag
-            url = reverse('tag-detail', kwargs={'slug': proper.name})
-            return HttpResponsePermanentRedirect(url)
-        else:
-            return super(TagDetail, self).get(*args, **kwargs)
-
-    def get_context_data(self, **kwargs):
-        context = super(TagDetail, self).get_context_data(**kwargs)
-        knesset = self.get_knesset()
-        context['knesset_id'] = knesset
-        tag = context['object']
-
-        cms_date_filter, proposal_date_filter, vote_date_filter = self._resolve_date_filters(knesset)
-
-        votes = Vote.objects.filter(vote_date_filter)
-        cms = CommitteeMeeting.objects.get_only_commitees().filter(cms_date_filter)
-
-        context['bills'] = self._get_relevant_bills_by_tag(cms, proposal_date_filter, tag, votes)
-
-        context['votes'] = self._get_relevant_votes_by_tag(tag, votes)
-
-        context['cms'] = self._get_relevant_committee_meetings_by_tag(cms, tag)
-
-        (context['members'],
-         context['past_members']) = self.create_tag_cloud(tag)
-
-        context['paginator'] = paginator = SelectorPaginator(Knesset.objects.values_list('number', flat=True))
-        context['page_obj'] = paginator.page(knesset.number)
-        context['request'] = None
-
-        return context
-
-    def _get_relevant_committee_meetings_by_tag(self, cms, tag):
-        cm_ct = ContentType.objects.get_for_model(CommitteeMeeting)
-        cm_ids = TaggedItem.objects.filter(
-            tag=tag, content_type=cm_ct).values_list('object_id', flat=True)
-        cms = cms.filter(id__in=cm_ids)
-        return cms
-
-    def _get_relevant_votes_by_tag(self, tag, votes):
-        votes_ct = ContentType.objects.get_for_model(Vote)
-        vote_ids = TaggedItem.objects.filter(
-            tag=tag, content_type=votes_ct).values_list('object_id', flat=True)
-        votes = votes.filter(id__in=vote_ids).order_by('-time')
-        return_all_committee_meetings = utils.parse_boolean(self.request.GET.get('all_committees', False))
-        if not return_all_committee_meetings:
-            votes = votes[:10]
-        return votes
-
-    def _get_relevant_bills_by_tag(self, cms, proposal_date_filter, tag, votes):
-        knesset_bills = Bill.objects.filter(
-            Q(pre_votes__in=votes)
-            | Q(first_committee_meetings__in=cms)
-            | Q(first_vote__in=votes)
-            | Q(second_committee_meetings__in=cms)
-            | Q(approval_vote__in=votes)
-            | proposal_date_filter
-        ).distinct()
-        bills_ct = ContentType.objects.get_for_model(Bill)
-        bill_ids = TaggedItem.objects.filter(
-            tag=tag,
-            content_type=bills_ct).values_list('object_id', flat=True)
-        bills = knesset_bills.filter(id__in=bill_ids)
-        return bills
-
-    def _resolve_date_filters(self, knesset):
-        knesset_end_date = knesset.end_date
-        knesset_start_date = knesset.start_date
-        if knesset_end_date is not None:
-            cms_date_filter = Q(date__gte=knesset_start_date, date__lte=knesset_end_date)
-            vote_date_filter = Q(time__gte=knesset_start_date, time__lte=knesset_end_date)
-            proposal_date_filter = Q(proposals__date__gte=knesset_start_date, proposals__date__lte=knesset_end_date)
-        else:
-            cms_date_filter = Q(date__gte=knesset_start_date)
-            vote_date_filter = Q(time__gte=knesset_start_date)
-            proposal_date_filter = Q(proposals__date__gte=knesset_start_date)
-        return cms_date_filter, proposal_date_filter, vote_date_filter
-
-    def get_knesset(self):
-        try:
-            needed_knesset_id = self.request.GET['page']
-            try:
-                knesset = Knesset.objects.filter(number=needed_knesset_id)[0]
-            except IndexError as e:
-                raise Http404('Invalid knesset number (page=%s)' % (needed_knesset_id,))
-        except KeyError as e:
-            knesset = Knesset.objects.current_knesset()
-        return knesset
-
-
-class CsvView(BaseListView):
-    """A view which generates CSV files with information for a model queryset.
-    Important class members to set when inheriting:
-      * model -- the model to display information from.
-      * queryset -- the query performed on the model; defaults to all.
-      * filename -- the name of the resulting CSV file (e.g., "info.csv").
-      * list_display - a list (or tuple) of tuples, where the first item in
-        each tuple is the attribute (or the method) to display and
-        the second item is the title of that column.
-
-        The attribute can be a attribute on the CsvView child or the model
-        instance itself. If it's a callable it'll be called with (obj, attr)
-        for the CsvView attribute or without params for the model attribute.
-    """
-
-    filename = None
-    list_display = None
-
-    def dispatch(self, request):
-        if None in (self.filename, self.list_display, self.model):
-            raise Http404()
-        self.request = request
-        response = HttpResponse(mimetype='text/csv')
-        response['Content-Disposition'] = \
-            'attachment; filename="{}"'.format(self.filename)
-
-        object_list = self.get_queryset()
-        self.prepare_csv_for_utf8(response)
-        writer = csv.writer(response, dialect='excel')
-        writer.writerow([title.encode('utf8')
-                         for _, title in self.list_display])
-        for obj in object_list:
-            row = [self.get_display_attr(obj, attr)
-                   for attr, _ in self.list_display]
-            writer.writerow([unicode(item).encode('utf8') for item in row])
-        return response
-
-    def get_display_attr(self, obj, attr):
-        """Return the display string for an attr, calling it if necessary."""
-        display_attr = getattr(self, attr, None)
-        if display_attr is not None:
-            if callable(display_attr):
-                display_attr = display_attr(obj, attr)
-        else:
-            display_attr = getattr(obj, attr)
-            if callable(display_attr):
-                display_attr = display_attr()
-        if display_attr is None:
-            return ""
-        return display_attr
-
-    @staticmethod
-    def prepare_csv_for_utf8(fileobj):
-        """Prepend a byte order mark (BOM) to a file.
-
-        When Excel opens a CSV file, it assumes the encoding is ASCII. The BOM
-        directs it to decode the file with utf-8.
-        """
-        fileobj.write('\xef\xbb\xbf')
-
-
-class GetMoreView(ListView):
-    """A base view for feeding data to 'get more...' type of links
-
-    Will return a json result, with partial of rendered template:
-    {
-        "content": "....",
-        "current": current_patge number
-        "total": total_pages
-        "has_next": true if next page exists
-    }
-    We'll paginate the response. Since Get More link targets may already have
-    initial data, we'll look for `initial` GET param, and take it into
-    consdiration, completing to page size.
-    """
-
-    def get_context_data(self, **kwargs):
-        ctx = super(GetMoreView, self).get_context_data(**kwargs)
-        try:
-            initial = int(self.request.GET.get('initial', '0'))
-        except ValueError:
-            initial = 0
-
-        # initial only affects on first page
-        if ctx['page_obj'].number > 1 or initial >= self.paginate_by - 1:
-            initial = 0
-
-        ctx['object_list'] = ctx['object_list'][initial:]
-        return ctx
-
-    def render_to_response(self, context, **response_kwargs):
-        """We'll take the rendered content, and shove it into json"""
-
-        tmpl_response = super(GetMoreView, self).render_to_response(
-            context, **response_kwargs).render()
-
-        page = context['page_obj']
-
-        result = {
-            'content': tmpl_response.content,
-            'total': context['paginator'].num_pages,
-            'current': page.number,
-            'has_next': page.has_next(),
-        }
-
-        return HttpResponse(json.dumps(result, ensure_ascii=False),
-                            content_type='application/json')
-
-
-def untagged_objects(request):
-    return render_to_response('auxiliary/untagged_objects.html', {
-        'cms': CommitteeMeeting.objects.filter_and_order(tagged=['false'])[:100],
-        'cms_count': CommitteeMeeting.objects.filter_and_order(tagged=['false']).count(),
-        'bills': Bill.objects.filter_and_order(tagged='false')[:100],
-        'bill_count': Bill.objects.filter_and_order(tagged='false').count(),
-        'votes': Vote.objects.filter_and_order(tagged='false')[:100],
-        'vote_count': Vote.objects.filter_and_order(tagged='false').count(),
-    },
-                              context_instance=RequestContext(request))
